added express router, cleaned up code
