+++ package.json
@@ -5,7 +5,7 @@
   "repository": "https://github.com/mtrew2015/auth-ii.git",
   "license": "MIT",
   "scripts": {
-    "server" : "nodemon"
+    "server": "nodemon"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
@@ -16,6 +16,7 @@
     "helmet": "^3.15.0",
     "jsonwebtoken": "^8.4.0",
     "knex": "^0.16.3",
+    "morgan": "^1.9.1",
     "sqlite3": "^4.0.6",
     "uuid": "^3.3.2"
   },
